Mapping of the shadow stack to one page.
The shadow stack is now mapped to a complete page. Hence BinProtect is
not only capable of supporting more active stack frames, but also allows
a simpler remapping of memory: BinProtect maps the memory required for
the shadow stack (as part of the .dyninst text/data segment -- initially
marked as RWX) as RW. In addition, BinProtect remaps the remaining part
of the .dyninst text/data segment as RE. This way, the attack vector is
reduced, since loaded memory cannot be written to, when it is marked as
executable.