Fixed bug in prolo and epilog.
The old implementation had a bug that did not consider the actual return
addresses but took another value on the stack to be protected. This has
been fixed.

In addition, the older implementation did not check for valid saved EIP
values. It assumed them to be valid, which sometimes led to crashes in
case the EBP value has not yet been set (see main function). This has
been fixed as well.